+++ packages/aws-amplify-angular/src/components/authenticator/index.ts
@@ -1,8 +1,8 @@
-export { AuthenticatorComponent } from './authenticator.component'
-export { SignInComponent } from './sign-in.component'
-export { RequireNewPasswordComponent } from './require-new-password.component'
-export { ConfirmSignInComponent } from './confirm-sign-in.component'
-export { SignUpComponent } from './sign-up.component'
-export { ConfirmSignUpComponent } from './confirm-sign-up.component'
-export { ForgotPasswordComponent } from './forgot-password.component'
-export { GreetingsComponent } from './greetings.component'
+export * from './authenticator';
+export * from './sign-in-component';
+export * from './sign-up-component';
+export * from './confirm-sign-up-component';
+export * from './confirm-sign-in-component';
+export * from './forgot-password-component';
+export * from './greeting-component';
+export * from './require-new-password-component';