+++ packages/aws-amplify-react-native/src/index.js
@@ -21,6 +21,7 @@ export * from './AmplifyUI';
 export * from './Auth';
 export * from './API';
 export * from './Storage';
+export * from './Interactions';
 export { default as PushNotification } from './PushNotification'
 
 const configure = function(config) {