+++ packages/aws-amplify/lib/Common/RNComponents/index.d.ts
@@ -1,5 +1,0 @@
-declare const Linking: {};
-declare const AppState: {
-    addEventListener(action: any, handler: any): void;
-};
-export { Linking, AppState };