+++ packages/aws-amplify-react-native/CHANGELOG.md
@@ -3,6 +3,178 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.20-unstable.1"></a>
+## [0.2.20-unstable.1](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.20-unstable.0...aws-amplify-react-native@0.2.20-unstable.1) (2018-07-03)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.20-unstable.0"></a>
+## [0.2.20-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.19...aws-amplify-react-native@0.2.20-unstable.0) (2018-07-02)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.19"></a>
+## [0.2.19](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.19-unstable.5...aws-amplify-react-native@0.2.19) (2018-06-29)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.19-unstable.5"></a>
+## [0.2.19-unstable.5](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.19-unstable.4...aws-amplify-react-native@0.2.19-unstable.5) (2018-06-29)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.19-unstable.4"></a>
+## [0.2.19-unstable.4](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.19-unstable.3...aws-amplify-react-native@0.2.19-unstable.4) (2018-06-29)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.19-unstable.3"></a>
+## [0.2.19-unstable.3](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.19-unstable.2...aws-amplify-react-native@0.2.19-unstable.3) (2018-06-28)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.19-unstable.2"></a>
+## [0.2.19-unstable.2](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.19-unstable.1...aws-amplify-react-native@0.2.19-unstable.2) (2018-06-27)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.19-unstable.1"></a>
+## [0.2.19-unstable.1](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.19-unstable.0...aws-amplify-react-native@0.2.19-unstable.1) (2018-06-27)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.19-unstable.0"></a>
+## [0.2.19-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.18-unstable.2...aws-amplify-react-native@0.2.19-unstable.0) (2018-06-27)
+
+
+### Features
+
+* **interactions:** Interactions UI components for react and react native ([#1105](https://github.com/aws/aws-amplify/issues/1105)) ([57de248](https://github.com/aws/aws-amplify/commit/57de248))
+
+
+
+
+<a name="0.2.18"></a>
+## [0.2.18](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.18-unstable.2...aws-amplify-react-native@0.2.18) (2018-06-27)
+
+
+### Features
+
+* **interactions:** Interactions UI components for react and react native ([#1105](https://github.com/aws/aws-amplify/issues/1105)) ([57de248](https://github.com/aws/aws-amplify/commit/57de248))
+
+
+
+
+<a name="0.2.18-unstable.2"></a>
+## [0.2.18-unstable.2](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.18-unstable.1...aws-amplify-react-native@0.2.18-unstable.2) (2018-06-26)
+
+
+### Bug Fixes
+
+* **integration tests:** CircleCI workflows and Cypress integration testing ([#1071](https://github.com/aws/aws-amplify/issues/1071)) ([bfa4776](https://github.com/aws/aws-amplify/commit/bfa4776))
+
+
+
+
+<a name="0.2.18-unstable.1"></a>
+## [0.2.18-unstable.1](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.18-unstable.0...aws-amplify-react-native@0.2.18-unstable.1) (2018-06-22)
+
+
+### Bug Fixes
+
+* **aws-amplify-react-native:** firebase-messaging and firebase-core version update ([a1031ec](https://github.com/aws/aws-amplify/commit/a1031ec))
+
+
+
+
+<a name="0.2.18-unstable.0"></a>
+## [0.2.18-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.17...aws-amplify-react-native@0.2.18-unstable.0) (2018-06-22)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.17"></a>
+## [0.2.17](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.16-unstable.3...aws-amplify-react-native@0.2.17) (2018-06-21)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.16"></a>
+## [0.2.16](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.13-unstable.29...aws-amplify-react-native@0.2.16) (2018-06-20)
+<a name="0.2.16-unstable.3"></a>
+## [0.2.16-unstable.3](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.16-unstable.2...aws-amplify-react-native@0.2.16-unstable.3) (2018-06-21)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.16-unstable.2"></a>
+## [0.2.16-unstable.2](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.16-unstable.1...aws-amplify-react-native@0.2.16-unstable.2) (2018-06-21)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.16-unstable.1"></a>
+## [0.2.16-unstable.1](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.16-unstable.0...aws-amplify-react-native@0.2.16-unstable.1) (2018-06-20)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="0.2.16-unstable.0"></a>
+## [0.2.16-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.13-unstable.29...aws-amplify-react-native@0.2.16-unstable.0) (2018-06-20)
+
+
+### Bug Fixes
+
+* **pushnotification:** revert change in pr 952 ([b8d167c](https://github.com/aws/aws-amplify/commit/b8d167c))
+* **pushnotification:** revert change in pr 952 ([257fc40](https://github.com/aws/aws-amplify/commit/257fc40))
+
+
+
+
+<a name="0.2.15"></a>
+## [0.2.15](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.13-unstable.15...aws-amplify-react-native@0.2.15) (2018-06-04)
+<a name="0.2.13-unstable.29"></a>
+## [0.2.13-unstable.29](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.13-unstable.28...aws-amplify-react-native@0.2.13-unstable.29) (2018-06-19)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
 <a name="0.2.13-unstable.28"></a>
 ## [0.2.13-unstable.28](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.13-unstable.27...aws-amplify-react-native@0.2.13-unstable.28) (2018-06-18)
 
@@ -113,19 +285,22 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Bug Fixes
 
-* **pushNotification:** revert the change in pr 952 ([2e03162](https://github.com/aws/aws-amplify/commit/2e03162))
+* **pushnotification:** revert change in pr 952 ([b8d167c](https://github.com/aws/aws-amplify/commit/b8d167c))
+* **pushnotification:** revert change in pr 952 ([257fc40](https://github.com/aws/aws-amplify/commit/257fc40))
 
 
 
 
-<a name="0.2.13-unstable.14"></a>
-## [0.2.13-unstable.14](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.13-unstable.13...aws-amplify-react-native@0.2.13-unstable.14) (2018-06-04)
+<a name="0.2.14"></a>
+## [0.2.14](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.13...aws-amplify-react-native@0.2.14) (2018-06-02)
 
 
 
 
 **Note:** Version bump only for package aws-amplify-react-native
 
+<a name="0.2.13"></a>
+## [0.2.13](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.13-unstable.12...aws-amplify-react-native@0.2.13) (2018-06-01)
 <a name="0.2.13-unstable.13"></a>
 ## [0.2.13-unstable.13](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@0.2.13-unstable.12...aws-amplify-react-native@0.2.13-unstable.13) (2018-06-02)